Add multiple references to nodes based on ports
Previously, nodes were referenced by their outpur wire from
the list of nodes in a CircuitGraph. This change makes it
so you can reference a node on three criteria:
 - output wire
 - output port
 - input ports

For the Verilog line:

  and AND1(w1, X, Y)

you can reference this gate by:
 - ckt_graph.nodes["w1"]
 - ckt_graph.nodes["AND1.Y"]

And you can reference nodes A and B by:
 - ckt_graph.nodes["AND1.A"]
 - ckt_graph.nodes["AND1.B"]

Nodes are ordered A, B, C, etc. based on the order they appear
in the Verilog.