Make checking key correctness optional
Checking if the key is correct is not a part of the SAT attack and
can cause issues when measuring the time it takes to perform a
SAT attack. I first noticed this when running a SAT attack against
optimized benchmarks. The optimized benchmarks should have had a
lower SAT attack time, but all of them were much higher. When I
manually ran the SAT attacks, I noticed that most of the time
was being spent checking for the correct key for some reason.
Therefore, I made this optional so I can accurately measure the
time it takes to run the SAT attack on the optimized benchmarks.